ディレクトリーの構造変更に対応

+++ scripts/merge_dic.sh
@@ -4,20 +4,18 @@
 SRC_DIR='.'
 
 # 同義表現データディレクトリ
-SIM_DIR_Dic=../dic/dic_rsk_iwanami
-SIM_DIR_Web=../dic/dic_web
+SIM_DIR_Dic=../dic/rsk_iwanami
+SIM_DIR_Web=../dic/web_news
 
 # 同義表現データマージ後のディレクトリ
 SIM_C_DIR=../dic_change
 
 # 同義表現データベース
-SYNDB_DIR=../syndb/i686
+SYNDB_DIR=../syndb/cgi
 
-while getopts oh OPT
+while getopts h OPT
 do
   case $OPT in
-      o)  SYNDB_DIR=../syndb/x86_64
-          ;;
       h)  usage
           ;;
     esac
@@ -33,15 +31,16 @@ export PERL5LIB=$SRC_DIR:$PERL5LIB
 
 
 # 全部削除する
-for f in definition.txt synonym_dic.txt isa.txt antonym.txt synonym_web.txt log_merge.txt; do
+for f in definition.txt synonym_dic.txt isa.txt antonym.txt synonym_web_news.txt log_merge.txt; do
     if [ -e $SIM_C_DIR/$f ] ; then
 	rm -v $SIM_C_DIR/$f
     fi
 done
 
-rm -v $SYNDB_DIR/log_dic.cdb
-#rm -v $SIM_C_DIR/definition.txt $SIM_C_DIR/synonym_rsk.txt $SIM_C_DIR/isa.txt $SIM_C_DIR/antonym.txt $SIM_C_DIR/synonym_web.txt 
-#rm -v $SIM_C_DIR/log_merge.txt
+if [ -e $SYNDB_DIR/log_dic.cdb ] ; then
+    rm -v $SYNDB_DIR/log_dic.cdb
+fi
+
 
 ########################################################
 echo "STEP1 start\t`date`"
@@ -55,5 +54,4 @@ perl -I$PERL_DIR change_dic.pl --synonym=$SIM_DIR_Dic/synonym.txt --definition=$
 cp $SIM_DIR_Dic/definition.txt $SIM_C_DIR/definition.txt
 
 # Webからの辞書の整理
-# perl check_duplicate_entry.pl -rnsame < $SIM_DIR/synonym_web.txt > $SIM_C_DIR/synonym_web.txt
-perl check_duplicate_entry.pl -rnsame --synonym_web=$SIM_DIR_Web/all.txt --log_merge=$SIM_C_DIR/log_merge.txt --change=$SIM_C_DIR/synonym_web.txt
+perl check_duplicate_entry.pl -rnsame --synonym_web_news=$SIM_DIR_Web/all.txt --log_merge=$SIM_C_DIR/log_merge.txt --change=$SIM_C_DIR/synonym_web_news.txt
